backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langgraph>=0.1.0
langchain>=0.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
sse-starlette>=2.1.0
pydantic>=2.0.0
